requirements.txt
discord.py==1.7.3
DBSkr==2.0
asyncio~=3.4.3